[ SEARCH-1429 ] Tests separation between /afts and !afts
