+++ alfresco-search/src/test/java/org/alfresco/solr/AbstractAlfrescoSolrTests.java
@@ -36,11 +36,13 @@ import javax.xml.xpath.XPathExpressionException;
 import org.alfresco.repo.search.impl.parsers.FTSQueryParser;
 import org.alfresco.service.cmr.repository.NodeRef;
 import org.alfresco.service.cmr.search.SearchParameters;
+import org.alfresco.service.namespace.QName;
 import org.alfresco.solr.client.Node;
 import org.alfresco.solr.client.NodeMetaData;
 import org.alfresco.solr.client.SOLRAPIQueueClient;
 import org.alfresco.solr.client.Transaction;
 import org.alfresco.solr.tracker.Tracker;
+import org.alfresco.util.SearchLanguageConversion;
 import org.apache.chemistry.opencmis.commons.impl.json.JSONArray;
 import org.apache.chemistry.opencmis.commons.impl.json.JSONObject;
 import org.apache.chemistry.opencmis.commons.impl.json.JSONValue;
@@ -81,6 +83,9 @@ import org.junit.AfterClass;
 import org.junit.Assert;
 import org.xml.sax.SAXException;
 
+import static java.util.Optional.ofNullable;
+import static org.junit.Assert.assertEquals;
+
 /**
  * Base class that provides the solr test harness.
  * This is used to manage the embedded solr used for unit and integration testing.
@@ -93,7 +98,7 @@ import org.xml.sax.SAXException;
 public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, AlfrescoSolrConstants
 {
     static AlfrescoCoreAdminHandler admin;
-    private static Log log = LogFactory.getLog(AbstractAlfrescoSolrTests.class);
+    private static Log LOG = LogFactory.getLog(AbstractAlfrescoSolrTests.class);
 
     protected static SolrConfig solrConfig;
     /**
@@ -114,11 +119,11 @@ public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, Alfre
      */
     protected static TestHarness.LocalRequestFactory lrf;
     protected static AlfrescoSolrDataModel dataModel = AlfrescoSolrDataModel.getInstance();
-    protected static NodeRef testRootNodeRef;
-    protected static NodeRef testNodeRef;
+    protected static NodeRef TEST_ROOT_NODEREF;
+    protected static NodeRef TEST_NODEREF;
     protected static NodeRef testBaseFolderNodeRef;
     protected static NodeRef testFolder00NodeRef;
-    protected static Date ftsTestDate;
+    protected static Date FTS_TEST_DATE;
     /**
      * Creates a Solr Alfresco test harness.
      * @param schema
@@ -126,8 +131,8 @@ public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, Alfre
      */
     public static void initAlfrescoCore(String schema) throws Exception
     {
-        log.info("##################################### init Alfresco core ##############");
-        log.info("####initCore");
+        LOG.info("##################################### init Alfresco core ##############");
+        LOG.info("####initCore");
         System.setProperty("solr.solr.home", TEST_FILES_LOCATION);
         System.setProperty("solr.directoryFactory","solr.RAMDirectoryFactory");
         System.setProperty("solr.tests.maxBufferedDocs", "1000");
@@ -143,7 +148,7 @@ public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, Alfre
         {
             createAlfrescoCore(schema);
         }
-        log.info("####initCore end");
+        LOG.info("####initCore end");
         admin = (AlfrescoCoreAdminHandler)h.getCore().getCoreContainer().getMultiCoreHandler();
     }
 
@@ -175,7 +180,7 @@ public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, Alfre
         SolrResourceLoader resourceLoader = new SolrResourceLoader(Paths.get(TEST_SOLR_CONF), null, properties);
         solrConfig = new SolrConfig(resourceLoader, "solrconfig.xml", null);
         IndexSchema indexSchema = IndexSchemaFactory.buildIndexSchema(schema, solrConfig);
-        log.info("################ Index schema:"+schema+":"+indexSchema.getResourceName());
+        LOG.info("################ Index schema:"+schema+":"+indexSchema.getResourceName());
         TestCoresLocator locator = new TestCoresLocator(SolrTestCaseJ4.DEFAULT_TEST_CORENAME,
                                                         "data", 
                                                         solrConfig.getResourceName(),
@@ -194,7 +199,7 @@ public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, Alfre
         }
         catch(Exception e)
         {
-            log.info("we hit an issue", e);
+            LOG.info("we hit an issue", e);
         }
         lrf = h.getRequestFactory
                 ("standard",0,20, CommonParams.VERSION,"2.2");
@@ -333,7 +338,7 @@ public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, Alfre
                 String msg = "REQUEST FAILED: xpath=" + results
                         + "\n\txml response was: " + response
                         + "\n\trequest was:" + req.getParamString();
-                log.error(msg);
+                LOG.error(msg);
                 throw new RuntimeException(msg);
             }
         } catch (XPathExpressionException e1) 
@@ -369,7 +374,7 @@ public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, Alfre
             {
                 JSONObject item = (JSONObject) doc;
                 BigInteger val = (BigInteger) item.get("DBID");
-                Assert.assertEquals(dbids[count].intValue(), val.intValue());
+                assertEquals(dbids[count].intValue(), val.intValue());
                 count++;
             }
         }
@@ -594,23 +599,23 @@ public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, Alfre
     {
         assertU(delQ("*:*"));
     }
-    protected void assertAQuery(String queryString,Integer count)throws IOException,ParseException 
+
+    protected void assertAQuery(String queryString,Integer count)
     {
         assertAQuery(queryString, count, null, null, null);
     }
 
-    protected void assertAQuery(String queryString,
-            Integer count,
+    protected void assertAQuery(
+            String queryString,
+            int count,
             Locale locale,
             String[] textAttributes,
             String[] allAttributes,
-            String... name)throws IOException,ParseException
-{
-        SolrServletRequest solrQueryRequest = null;
+            String... name)
+    {
         RefCounted<SolrIndexSearcher>refCounted = null;
-        try
+        try (SolrServletRequest solrQueryRequest = new SolrServletRequest(h.getCore(), null))
         {
-            solrQueryRequest = new SolrServletRequest(h.getCore(), null);
             refCounted = h.getCore().getSearcher();
             SolrIndexSearcher solrIndexSearcher = refCounted.get();
             SearchParameters searchParameters = new SearchParameters();
@@ -619,41 +624,42 @@ public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, Alfre
             {
                 searchParameters.addLocale(locale);
             }
-            
+
             if (textAttributes != null)
             {
                 for (String textAttribute : textAttributes)
-            {
+                {
                     searchParameters.addTextAttribute(textAttribute);
+                }
             }
-        }
-        if (allAttributes != null)
-        {
-            for (String allAttribute : allAttributes)
-            {
-                searchParameters.addAllAttribute(allAttribute);
-            }
-        }
-        
-        Query query = dataModel.getLuceneQueryParser(searchParameters, solrQueryRequest, FTSQueryParser.RerankPhase.SINGLE_PASS).parse(queryString);
-        log.debug("####### Query ######:"+query);
-        TopDocs docs = solrIndexSearcher.search(query, count * 2 + 10);
 
-        if (count != null)
-        {
-            if (docs.totalHits != count)
+            if (allAttributes != null)
             {
-                throw new IOException("FAILED: " + fixQueryString(queryString, name)+" ; "+docs.totalHits);
+                for (String allAttribute : allAttributes)
+                {
+                    searchParameters.addAllAttribute(allAttribute);
+                }
             }
+
+            Query query = dataModel.getLuceneQueryParser(searchParameters, solrQueryRequest, FTSQueryParser.RerankPhase.SINGLE_PASS).parse(queryString);
+
+            LOG.debug("Query: " + query);
+
+            TopDocs docs = solrIndexSearcher.search(query, count * 2 + 10);
+
+            assertEquals(fixQueryString(queryString, name), count, docs.totalHits);
         }
-    }
+        catch(Exception exception)
+        {
+            throw new RuntimeException(exception);
+        }
         finally
         {
-            refCounted.decref();
-            solrQueryRequest.close();
+            ofNullable(refCounted).ifPresent(RefCounted::decref);
         }
     }
-    protected String fixQueryString(String queryString, String... name)
+
+    private String fixQueryString(String queryString, String... name)
     {
         if (name.length > 0)
         {
@@ -664,6 +670,7 @@ public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, Alfre
             return queryString.replace("\uFFFF", "<Unicode FFFF>");
         }
     }
+
     /**
      * Generates a SolrQueryRequest
      */
@@ -686,6 +693,7 @@ public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, Alfre
         }
         return msp;
     }
+
     public static class SolrServletRequest extends SolrQueryRequestBase
     {
         public SolrServletRequest(SolrCore core, HttpServletRequest req)
@@ -693,9 +701,20 @@ public abstract class  AbstractAlfrescoSolrTests implements SolrTestFiles, Alfre
             super(core, new MultiMapSolrParams(Collections.<String, String[]> emptyMap()));
         }
     }
+
     protected Collection<Tracker> getTrackers() {
         Collection<Tracker> trackers = admin.getTrackerRegistry().getTrackersForCore(h.getCore().getName());
-        log.info("######### Number of trackers is "+trackers.size()+" ###########");
+        LOG.info("######### Number of trackers is "+trackers.size()+" ###########");
         return trackers;
+    }
+
+    protected String escape(QName qname)
+    {
+        return escape(qname.toString());
+    }
+
+    protected String escape(String value)
+    {
+        return SearchLanguageConversion.escapeLuceneQuery(value);
     }
 }
